requirements.txt
langchain==0.1.16
langchain_community==0.0.34
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.32.2